Added final polish
Now you can keep the kickers up while holding the keys, the horse plays
animation and sound, we have a new music theme and added new fx for the
tunnel and the cart and for when a new game begins. You can also look
how much your points are getting multiplied for.